requirements.txt
streamlit
requests
ffmpeg-python
transformers
torch --pre --index-url https://download.pytorch.org/whl/nightly/cu128
fastapi
uvicorn
numpy
soundfile
